+++ misc/build_scripts/windows/exiftool.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-EXIFTOOL_TAG="13.25"
+EXIFTOOL_TAG="13.30"
 
 RUN_DIR=$(pwd)
 OUTPUT_DIR="${RUN_DIR}/bin/win/exiftool"